# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vybe - Group Dining Decider API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface GoogleAuthRequest {
  idToken: string;
  accessToken?: string;
}

export interface GuestAuthRequest {
  name: string;
  avatar?: string;
}

export type UserProvider = (typeof UserProvider)[keyof typeof UserProvider];

export const UserProvider = {
  google: "google",
  guest: "guest",
} as const;

export interface User {
  id: string;
  name: string;
  email?: string;
  avatar?: string;
  provider: UserProvider;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type CreateGroupRequestVenueType =
  (typeof CreateGroupRequestVenueType)[keyof typeof CreateGroupRequestVenueType];

export const CreateGroupRequestVenueType = {
  restaurant: "restaurant",
  cafe: "cafe",
  bar: "bar",
} as const;

export interface CreateGroupRequest {
  name: string;
  memberIds: string[];
  venueType: CreateGroupRequestVenueType;
}

export type GroupVenueType =
  (typeof GroupVenueType)[keyof typeof GroupVenueType];

export const GroupVenueType = {
  restaurant: "restaurant",
  cafe: "cafe",
  bar: "bar",
} as const;

export type GroupStatus = (typeof GroupStatus)[keyof typeof GroupStatus];

export const GroupStatus = {
  pending: "pending",
  swiping: "swiping",
  waiting: "waiting",
  completed: "completed",
} as const;

export type GroupMemberStatus =
  (typeof GroupMemberStatus)[keyof typeof GroupMemberStatus];

export const GroupMemberStatus = {
  invited: "invited",
  joined: "joined",
  done: "done",
} as const;

export interface GroupMember {
  userId: string;
  name: string;
  avatar?: string;
  status: GroupMemberStatus;
  doneAt?: string;
}

export interface Group {
  id: string;
  name: string;
  creatorId: string;
  venueType: GroupVenueType;
  status: GroupStatus;
  members: GroupMember[];
  restaurantCount?: number;
  createdAt?: string;
}

export interface StartSwipingRequest {
  lat: number;
  lng: number;
  radius: number;
  locationName?: string;
}

export type MenuItemCategory =
  (typeof MenuItemCategory)[keyof typeof MenuItemCategory];

export const MenuItemCategory = {
  dish: "dish",
  drink: "drink",
  dessert: "dessert",
} as const;

export interface MenuItem {
  name: string;
  description?: string;
  category: MenuItemCategory;
  price?: string;
  photo?: string;
}

export interface Review {
  authorName: string;
  rating: number;
  text: string;
  timeAgo?: string;
  authorPhoto?: string;
}

export interface Restaurant {
  id: string;
  placeId?: string;
  name: string;
  address: string;
  rating?: number;
  reviewCount?: number;
  priceLevel?: number;
  photos?: string[];
  cuisine?: string;
  openNow?: boolean;
  distance?: number;
  menuItems?: MenuItem[];
  reviews?: Review[];
  lat?: number;
  lng?: number;
  mapsUrl?: string;
}

export interface VoteRequest {
  restaurantId: string;
  liked: boolean;
}

export interface VoteResponse {
  allDone: boolean;
  doneCount: number;
  totalCount: number;
}

export interface RestaurantResult {
  restaurant: Restaurant;
  likeCount: number;
  dislikeCount: number;
  totalVotes: number;
  score: number;
}

export type SearchUsersParams = {
  q: string;
};

export type SearchRestaurantsParams = {
  lat: number;
  lng: number;
  radius: number;
  type?: SearchRestaurantsType;
};

export type SearchRestaurantsType =
  (typeof SearchRestaurantsType)[keyof typeof SearchRestaurantsType];

export const SearchRestaurantsType = {
  restaurant: "restaurant",
  cafe: "cafe",
  bar: "bar",
} as const;