# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vybe - Group Dining Decider API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/google:
    post:
      operationId: authGoogle
      tags: [auth]
      summary: Authenticate with Google token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleAuthRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/guest:
    post:
      operationId: authGuest
      tags: [auth]
      summary: Create guest session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestAuthRequest"
      responses:
        "200":
          description: Guest session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by name or email
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    GoogleAuthRequest:
      type: object
      properties:
        idToken:
          type: string
        accessToken:
          type: string
      required: [idToken]
    GuestAuthRequest:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: string
      required: [name]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar:
          type: string
        provider:
          type: string
          enum: [google, guest]
      required: [id, name, provider]
...